extract CompletionManager and QueryEditorCompletionData to component proj
